Renaming tests to .tsx
