+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  "transform": {
+    "^.+\\.tsx?$": `ts-jest`,
+    "^.+\\.js?$": `babel-jest`,
+  },
+  "testRegex": `.test.(js?|jsx?|tsx?)$`,
+  "moduleFileExtensions": [
+    `ts`,
+    `tsx`,
+    `js`,
+    `jsx`,
+    `json`,
+    `node`
+  ],
+};
