+++ src/components/main-page-empty/main-page-empty.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import * as React from 'react';
 import renderer from 'react-test-renderer';
-import MainPageEmpty from '../main-page-empty/main-page-empty';
+import MainPageEmpty from './main-page-empty';
 
 describe(`MainPageEmpty`, () => {
   it(`renders correctly`, () => {
