+++ src/components/main-page/main-page.test.tsx
@@ -1,11 +1,12 @@
-import React from 'react';
+import * as React from 'react';
 import renderer from 'react-test-renderer';
 import {BrowserRouter} from 'react-router-dom';
+import leafletMock from '../../mocks/leaflet-mock';
 import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 import configureMockStore from 'redux-mock-store';
 
-import PlaceList from '../place-list/place-list';
+import {MainPage} from './main-page';
 import {Operation} from '../../reducer/data/data';
 import {Operation as OperationUser} from '../../reducer/user/user';
 import NameSpace from '../../reducer/name-space';
@@ -16,20 +17,20 @@ Operation.addToFavorites = () => (dispatch) => dispatch(jest.fn());
 Operation.deleteFromFavorites = () => (dispatch) => dispatch(jest.fn());
 OperationUser.authorizeUser = () => (dispatch) => dispatch(jest.fn());
 
-describe(`PlaceList`, () => {
+describe(`MainPage`, () => {
   const places = [
     {
       id: 1,
       title: `Strange place`,
       isPremium: true,
       price: 1200,
-      rating: 1.8,
+      rating: 1.5,
       isFavorite: false,
       description: ``,
       type: `Apartment`,
       previewImage: ``,
-      images: [],
-      goods: [],
+      images: [``],
+      goods: [``],
       bedrooms: 2,
       maxAdults: 4,
       host: {},
@@ -53,12 +54,12 @@ describe(`PlaceList`, () => {
       isPremium: true,
       price: 800,
       rating: 1.5,
-      isFavorite: true,
+      isFavorite: false,
       description: ``,
       type: `Private room`,
       previewImage: ``,
-      images: [],
-      goods: [],
+      images: [``],
+      goods: [``],
       bedrooms: 2,
       maxAdults: 4,
       host: {},
@@ -122,12 +123,30 @@ describe(`PlaceList`, () => {
     isAuthorizationRequired: false,
   };
   const store = mockStore(initialState);
+  const cities = [`Berlin`, `Dusseldorf`];
 
   it(`renders correctly`, () => {
-    const tree = renderer.create(<BrowserRouter><Provider store={store}><PlaceList
-      offers={places}
-      onPlaceClick={jest.fn()}
-    /></Provider></BrowserRouter>).toJSON();
+    const tree = renderer.create(<BrowserRouter><Provider store={store}>
+      <MainPage
+        offers={places}
+        cities={cities}
+        city={{
+          name: `Dusseldorf`,
+          location: {
+            atitude: 52,
+            longitude: 8,
+            zoom: 11,
+          },
+        }}
+        onCityClick={jest.fn()}
+        leaflet={leafletMock}
+        onPlaceClick={jest.fn()}
+        activeCard={{}}
+        sortedOffers={places}
+        onSortingClick={jest.fn()}
+        activeSorting={1}
+      />
+    </Provider></BrowserRouter>).toJSON();
 
     expect(tree).toMatchSnapshot();
   });
