+++ src/components/place/place.e2e.test.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {Place} from '../place/place';
+import {Place} from './place';
 
 Enzyme.configure({adapter: new Adapter()});
 const offers = [
