+++ src/components/send-review-form/send-review-form.test.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import renderer from 'react-test-renderer';
 
-import {SendReviewForm} from './send-review-form/';
+import {SendReviewForm} from './send-review-form';
 
 describe(`SendReviewForm`, () => {
   it(`renders correctly`, () => {
