+++ src/hocs/with-active-card/with-active-card.e2e.test.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import withActiveCard from './with-active-card';
